requirements.txt
langchain
openai
streamlit
gradio
python-dotenv
